replace Broker.Users..Password with SecretKeyRef
Replaces the string `Password` field on the nested `Broker.Spec.Users`
field with a SecretKeyReference so that plain text passwords are not
passed in Broker CRDs.

This commit introduces a manual change to the
config/rbac/cluster-role-controller.yaml file to allow the controller to
read Secrets from the k8s API server. See
aws-controllers-k8s/community#745 for why that was necessary.

Issue aws-controllers-k8s/community#743