handle creation and deletion in progress
Adds hook code to the Broker resource manager that handles calls to
Update when a creation or deletion is in progress. We now requeue the
resource when attempting to update a resource that is in
CREATION_IN_PROGRESS state or when attempting to delete a resource that
is in DELETION_IN_PROGRESS state, which eliminates unnecessary error
logs in the controller.