Add `reservedConcurrentExecutions` field for `Function` Resource (#12)
The AWS Lambda API support setting up reserved concurrent executions for
a lambda functions. This patch adds a new field to allow users to set the
concurrency configuration for functions.

Description of changes:
- Add `reservedConcurrentExecutions` fields for `Function` Resource
- Add a new e2e test case to specifically test update and create
operations on function resources with a non-nil
`reservedConcurrentExecutions`

By submitting this pull request, I confirm that my contribution is made under
the terms of the Apache 2.0 license.