Support `functions CodeSigningConfig` updates
This patch essentially adds support for updating codeSigningConfig for
lambda functions. This patch also fixes a bug where the controllers
tries to update afunction that is in a Pending state - the fix is done
by requeing whenever a Status.State is 'Pending'. Along the way it fixes
few existing issues with the current e2e tests.

Description of changes:
- Requeue whenever Function.Status.State is Pending
- Do not try to update function code if only code.s3Key is modified
- Add a hook to prevent the create code path from trying to delete the
code signing config when it's given as an empty string
- Fixes a bug where created dynamodb tables were not deleted
- Fixes a bug where the cr used in the e2e tests were not completly up
to date and need to be refetched