Support functions `CodeSigningConfig` updates (#15)
Issue #, if available:

This patch essentially adds support for updating `CodeSigningConfig` for
lambda functions. This patch also fixes a bug where the controller
tries to update a function that is in a Pending state - the fix is done
by re-queuing whenever a `Status.State` is 'Pending'. Along the way, this
this patch also fixes few existing issues with the current e2e tests.

Description of changes:
- Requeue whenever `Function.Status.State` is `Pending`
- Do not try to update function code if only `code.s3Key` is modified
- Add a hook to prevent the create code path from trying to delete the
code signing config when it's given as an empty string
- Fixes a bug where created dynamodb tables were not deleted
- Fixes a bug where the cr used in the e2e tests were not completely up
to date and need to be refetched

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.