Stop defaulting non-modified fields when trying to update a Function
Fixes aws-controllers-k8s/community#1413

This patch modifies the update code path for a Function resource to
fix a few bugs observed while updating functions created using a
container image.

Description of the changes:

Only try to update Spec.Code or Spec.Configuration at once. It is
not correct to sequentially call UpdateFunctionConfiguration and
UpdateFunctionCode because both of them can put the function in a
Pending state.
Stop defaulting non-modified fields when trying to update a function.
Signed-off-by: Amine Hilaly hilalyamine@gmail.com

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>