Keep the desired `code.s3*` fields part of the spec
The AWS Lambda API doesn't return the original submitted `s3Bucket`,
`s3Key` and `s3ObjectVersion` when a function is successfully created.
Instead, it returns different values (In `Getfunction` Call) in a
different response field (`Code.Location`). This is causing the
generated code in ACK to set the `spec.Code.s3*` fields to nil values.

To keep the experience in ACK lambda-controller consistent with the
kubernetes exprience, we must keep the original submitted `code.S3*`
fields part of the function's spec.

This patch removes the part that is setting the `spec.Code.s3` to nil
and allows the controllers to keep the original values part of the spec.

Signed-off-by: Amine Hilaly <hilalyamine@gmail.com>