Adding delete all layer versions functionality for layerVersion resource (#70)
Description of changes:
Currently when you delete a LayerVersion it deletes the most recent version of it and keep the previous versions. The functionality we want is to delete all the versions of a LayerVersion when DeleteLayerVersion is executed.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.