Source image error handling (#98)
**Issue #, if available:**
https://github.com/aws-controllers-k8s/community/issues/1842

**Description of changes:**
Requeues the function reconciliation if it errors due to an image not existing. It will attempt reconciliation again after 1 minute. This allows for handling common GitOps conditions that may occur, where the function manifest is updated to use a new image URI, but the image hasn't finished pushing yet.

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.