Bug fix for `EventInvokeConfig`  (#107)
Issue #, if available:
Related PR: #92 #93 

**Summary**
Fixing deletion part for `EventInvokeConfig` field.

**Description:**
This PR is related to bug fixes in [previous PR](https://github.com/aws-controllers-k8s/lambda-controller/pull/97) related to `EventInvokeConfig` implementation. When the user deleted the set configurations for Function/Alias resource, the configurations were not deleted from the Console side, instead it used to throw an error. Fixed the bug by introducing a new condition to check if the desired configurations are nil or not, if yes then call the delete API. 

Example of changes for `EventInvokeConfig` in Alias resource is given below:

```
if r.ko.Spec.FunctionEventInvokeConfig == nil {
	input_delete := &svcsdk.DeleteFunctionEventInvokeConfigInput{
		FunctionName: aws.String(*r.ko.Spec.FunctionName),
		Qualifier:    aws.String(*r.ko.Spec.Name),
	}
	_, err = rm.sdkapi.DeleteFunctionEventInvokeConfigWithContext(ctx, input_delete)
	rm.metrics.RecordAPICall("DELETE", "DeleteFunctionEventInvokeConfig", err)
	if err != nil {
		return nil, err
	}
	return r, nil
}
```

This PR contains implementation code and E2E test to validate the changes. 

**Acknowledgment**
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.