fix: Support to update `architectures` field (#132)
**Issue:**
[#1971](https://github.com/aws-controllers-k8s/community/issues/1971)

**Description:**
When updating `architectures` field for the function code, it was not recognized by the lambda-controller. This PR fixes that bug. Now controller recognizes the change in `architectures` field and updates it.

**Acknowledgement:**
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.