Introduce `spec.code.sha256` field to help with s3 code changes (#88)
**Issue #**
[#1550](https://github.com/aws-controllers-k8s/community/issues/1550)

**Description**
This PR adds the feature to update the `Spec.Code.S3` variables. The issue arises as Lambda API does not keep record of `Spec.Code` variables, because of which ACK is not being able to compare the changes in the state for `Spec.Code` (ex: change in `Spec.Code.S3Key`) and thus fails to recognize the changes.

Thus we have added a new field `Spec.Code.SHA256`, the user has to manually calculate the SHA256 for their code, they can do this by running this command `sha256sum filename.zip |cut -f1 -d\ | xxd -r -p | base64`. This field will then be compared with current SHA256 taken from `Status.CodeSHA256` and will determine if the update has been made for deployment package. 

**Acknowledgment**
By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.