fix: ensure function syncs when inactive (#167)
Issue [#2446](https://github.com/aws-controllers-k8s/community/issues/2446)

Description of changes:
Currently the controller only marks a function resource as synced only 
when it is active. We want to ensure we sync if it is inactive as well.

In Lambda function, inactive simply indicates that a function has been idle
for a long time, and we do not want the controller to keep requeueing 
indefinitely

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.